+++ tests/EquivalenceTest.php
@@ -32,6 +32,7 @@ use Webmozart\Expression\Expression;
 use Webmozart\Expression\Logic\Not;
 use Webmozart\Expression\Selector\All;
 use Webmozart\Expression\Selector\AtLeast;
+use Webmozart\Expression\Selector\AtMost;
 use Webmozart\Expression\Selector\Exactly;
 use Webmozart\Expression\Selector\Key;
 
@@ -97,6 +98,8 @@ class EquivalenceTest extends PHPUnit_Framework_TestCase
             array(new Key('42', new Same('10')), new Key(42, new Same('10'))),
 
             array(new AtLeast(1, new Same('10')), new AtLeast(1, new Same('10'))),
+
+            array(new AtMost(1, new Same('10')), new AtMost(1, new Same('10'))),
 
             array(new Exactly(1, new Same('10')), new Exactly(1, new Same('10'))),
 
@@ -178,6 +181,10 @@ class EquivalenceTest extends PHPUnit_Framework_TestCase
             array(new AtLeast(1, new Same('10')), new AtLeast(2, new Same('10'))),
             array(new AtLeast(1, new Same('10')), new AtLeast(1, new Same(10))),
             array(new AtLeast(1, new Same('10')), new Same('10')),
+
+            array(new AtMost(1, new Same('10')), new AtMost(2, new Same('10'))),
+            array(new AtMost(1, new Same('10')), new AtMost(1, new Same(10))),
+            array(new AtMost(1, new Same('10')), new Same('10')),
 
             array(new Exactly(1, new Same('10')), new Exactly(2, new Same('10'))),
             array(new Exactly(1, new Same('10')), new Exactly(1, new Same(10))),